chore(sv): fix flaky test_search_with_region
